Move impl into cpp and not inlineing codec functions
